Fix for number

+++ Ubiq.FourcastersAPI/Models.cs
@@ -518,7 +518,7 @@ namespace Ubiq.FourcastersAPI
         public decimal odds { get; set; }
         public decimal amount { get; set; }
         public string type { get; set; }
-        public int? number { get; set; }
+        public decimal? number { get; set; }
         public string side { get; set; }
         public string wagerRequestID { get; set; }
         public string userReference { get; set; }
@@ -582,9 +582,9 @@ namespace Ubiq.FourcastersAPI
     {
         public string orderID { get; set; }
         public decimal offered { get; set; }
-        public int odds { get; set; }
+        public decimal odds { get; set; }
         public string type { get; set; }
-        public int? number { get; set; }
+        public decimal? number { get; set; }
         public string side { get; set; }
         public string wagerRequestID { get; set; }
         public string userReference { get; set; }
